+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 ## Fixed Issues
 
 - [core] Disable destination cache, when the JWT does not contain necessary information. For example, when using `IsolationStrategy.Tenant_User`, the JWT has to contain both tenant id and user id.
+- [core] Provider token used in retrieving destination from cache.
 
 
 # 1.51.0